fix: reorganize ec2 instance modules
